LayoutTests/accessibility/rowspan-zero-table-expected.txt

This test ensures we handle rowspan='0' tables correctly.

#table cellForColumnAndRow(0, 0).domIdentifier is r0c0
#table cellForColumnAndRow(1, 0).domIdentifier is r0c1
#table cellForColumnAndRow(2, 0).domIdentifier is r0c2
#table cellForColumnAndRow(0, 1).domIdentifier is r1c0-rowspan0-cell
#table cellForColumnAndRow(1, 1).domIdentifier is r1c1
#table cellForColumnAndRow(2, 1).domIdentifier is r1c2
#table cellForColumnAndRow(0, 2).domIdentifier is r1c0-rowspan0-cell
#table cellForColumnAndRow(1, 2).domIdentifier is r2c0
#table cellForColumnAndRow(2, 2).domIdentifier is r2c1
#table cellForColumnAndRow(0, 3).domIdentifier is r1c0-rowspan0-cell
#table cellForColumnAndRow(1, 3).domIdentifier is r3c0
#table cellForColumnAndRow(2, 3).domIdentifier is r3c1


{#table AXRole: AXTable}

{#r0 AXRole: AXRow}

{#r0c0 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Author}

{#r0c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Title}

{#r0c2 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Year}

{#r1 AXRole: AXRow}

{#r1c0-rowspan0-cell AXRole: AXCell}

{AXRole: AXStaticText AXValue: Stephen Hawking}

{#r1c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: A Brief History of Time}

{#r1c2 AXRole: AXCell}

{AXRole: AXStaticText AXValue: 1988}

{#r2 AXRole: AXRow}

{#r2c0 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Black Holes The Reith Lectures}

{#r2c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: 2016}

{#r3 AXRole: AXRow}

{#r3c0 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Brief Answers to the Big Questions}

{#r3c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: 2018}

{#r4c0 AXRole: AXGroup}

{AXRole: AXStaticText AXValue: A Briefer History of Time}

{#r4c1 AXRole: AXGroup}

{AXRole: AXStaticText AXValue: 2005}

Dynamically adding a new row.

PASS: table.rowCount === 5
#table cellForColumnAndRow(0, 0).domIdentifier is r0c0
#table cellForColumnAndRow(1, 0).domIdentifier is r0c1
#table cellForColumnAndRow(2, 0).domIdentifier is r0c2
#table cellForColumnAndRow(0, 1).domIdentifier is r1c0-rowspan0-cell
#table cellForColumnAndRow(1, 1).domIdentifier is r1c1
#table cellForColumnAndRow(2, 1).domIdentifier is r1c2
#table cellForColumnAndRow(0, 2).domIdentifier is r1c0-rowspan0-cell
#table cellForColumnAndRow(1, 2).domIdentifier is r2c0
#table cellForColumnAndRow(2, 2).domIdentifier is r2c1
#table cellForColumnAndRow(0, 3).domIdentifier is r1c0-rowspan0-cell
#table cellForColumnAndRow(1, 3).domIdentifier is r3c0
#table cellForColumnAndRow(2, 3).domIdentifier is r3c1
#table cellForColumnAndRow(0, 4).domIdentifier is r1c0-rowspan0-cell
#table cellForColumnAndRow(1, 4).domIdentifier is r4c0
#table cellForColumnAndRow(2, 4).domIdentifier is r4c1


{#table AXRole: AXTable}

{#r0 AXRole: AXRow}

{#r0c0 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Author}

{#r0c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Title}

{#r0c2 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Year}

{#r1 AXRole: AXRow}

{#r1c0-rowspan0-cell AXRole: AXCell}

{AXRole: AXStaticText AXValue: Stephen Hawking}

{#r1c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: A Brief History of Time}

{#r1c2 AXRole: AXCell}

{AXRole: AXStaticText AXValue: 1988}

{#r2 AXRole: AXRow}

{#r2c0 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Black Holes The Reith Lectures}

{#r2c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: 2016}

{#r3 AXRole: AXRow}

{#r3c0 AXRole: AXCell}

{AXRole: AXStaticText AXValue: Brief Answers to the Big Questions}

{#r3c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: 2018}

{#r4 AXRole: AXRow}

{#r4c0 AXRole: AXCell}

{AXRole: AXStaticText AXValue: A Briefer History of Time}

{#r4c1 AXRole: AXCell}

{AXRole: AXStaticText AXValue: 2005}


PASS successfullyParsed is true

TEST COMPLETE
This is a table caption
Author	Title	Year
Stephen Hawking	A Brief History of Time	1988
Black Holes The Reith Lectures	2016
Brief Answers to the Big Questions	2018
A Briefer History of Time	2005
